+++ 2022/day21/day21_test.go
@@ -77,7 +77,7 @@ func TestPart2(t *testing.T) { // nolint: dupl
 
 func TestRealData(t *testing.T) {
 	expected := `Result of part 1: 49288254556480
-Result of part 2: 1
+Result of part 2: 3558714869436
 `
 	buff := new(bytes.Buffer)
